requirements.txt
asgiref==3.8.1
Django==4.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
python-environ==0.4.54
sqlparse==0.5.0
tzdata==2024.1
